Thu, Sep 15, 2022, 4:54 PM +00:00

+++ readability/readability.c
@@ -55,7 +55,7 @@ int count_words(string text)
 				words++;
 			}
 		}
-		else ((int) text[0] == 32)
+		else (text[0] == 32)
 		{
 			break;
 		}
